source-sqlserver: Split stub methods into separate files
